feat(resourcing): enrich MCP read as get_resourcing_world (default_role_id, managed, degraded)
Renames GetResourcingProjectionsTool to GetResourcingWorldTool (tool_name
get_resourcing_world), adding people[].default_role_id (most-recent role by
endDate, from unfiltered history), a managed[] overlay of ProjectedAssignment
rows, and degraded/stale flags. Runn reads are wrapped in safe_read so a
single upstream failure degrades the payload instead of erroring the whole
call. get_resourcing_projections is kept as a thin deprecated subclass so
existing stacksbot prompts keep working.

Split into two files (get_resourcing_world_tool.rb + get_resourcing_projections_tool.rb)
rather than one, per Zeitwerk's one-file-one-constant convention — a single
file defining both classes made GetResourcingWorldTool unresolvable via
autoloading (only the file's expected constant, GetResourcingProjectionsTool,
was reachable), breaking every MCP tool call including unrelated ones.
